Implement getters inside ShopBuilderFactory and to be in the shop package.
